refactor helpers to dashcontroller
